fix(cpu_binding): handle localized npu-smi output
Remove English-only parsing assumptions from CPU binding by matching process rows structurally and extracting PCI addresses with a regex. Add regression tests for Chinese npu-smi headers and board output.

Also extract Ascend device type lookup into a lightweight module and add a logger fallback so the CPU binding unit tests can run in a minimal environment without the full torch_npu and vllm stack.

Fixes #6992
